requirements.txt
streamlit==1.31.0
pandas
requests
plotly.express
lxml
geopandas
html5lib
